Easier way to set the window icon
It's now possible to set the client's icon from a base64 encoded version of the imge! This way you don't need to include the .png image in the assets.
Update the Minecraft class (copy all of the three #setWindowIcon() methods) to get this feature.
For an example you can look at the BaseClient class