New Friends Manager
This is a better solution to manage friends and enemies.
In the previous "friends manager", if it can be called that, lowercase and uppercase would make a difference, whereas that has been fixed here, with a full actual friends manager.
This PR also updates the friends module to use the new friends manager.