Fix for XRay in other languages. (#46)
* Fix for XRay in other languages.

The XRay exception list was working with the language who is bad because if you change your language, you will need to add all the exceptions again in that language.
Now, it takes the English localized name whatever your language is.

* Updated to work on 1.8

I've loaded the english locale and put it in a variable available in the BaseClient class as I've anticipated before. I've tested it and it seems to be working just fine.
I will merge the PR tomorrow if no issue comes up before then

Co-authored-by: OxideWaveLength <60556091+OxideWaveLength@users.noreply.github.com>